Dodane CRUD funkcionalnosti za dnevnik te specijalizanta i specijalizaciju.
- funkcije za dodavanje, ažuriranje i brisanje svih sadržaja dnevnika (stručni rad, sučaj bolesnika, dnevna aktivnost, provjera znanja i pitanja) te specijalizanta i specijalizaciju

+++ Backend/ednevnik/src/main/java/hr/foi/air/ednevnik/Services/StrucniRadService.java
@@ -32,16 +32,13 @@ public class StrucniRadService {
     }
 
     public Long DeleteStrucniRad(int id) {
-        if(strucniRadRepository.existsById(id))
-        { return strucniRadRepository.deleteByIdRad(id); }
+        if(strucniRadRepository.existsById(id)) { return strucniRadRepository.deleteByIdRad(id); }
         else { return 0L; }
-
     }
 
     public StrucniRad UpdateStrucniRad(StrucniRad updatedStrucniRad) {
         Optional<StrucniRad> strucniRadZaUpdate = StrucniRadById(updatedStrucniRad.getIdRad());
         if(strucniRadZaUpdate.isEmpty()) { return null; }
         else { return strucniRadRepository.save(updatedStrucniRad); }
-
     }
 }
